moved Sink impl for Publisher to client
